+++ lib/flutter_typeahead.dart
@@ -659,6 +659,11 @@ class _TypeAheadFieldState<T> extends State<TypeAheadField<T>>
 
   final LayerLink _layerLink = LayerLink();
 
+  // Timer that resizes the suggestion box on each tick. Only active when the user is scrolling.
+  Timer _resizeOnScrollTimer;
+  // The rate at which the suggestion box will resize when the user is scrolling
+  final Duration _resizeOnScrollRefreshRate = const Duration(milliseconds: 500);
+
   @override
   void didChangeMetrics() {
     // Catch keyboard event and orientation change; resize suggestions list
@@ -669,6 +674,8 @@ class _TypeAheadFieldState<T> extends State<TypeAheadField<T>>
   void dispose() {
     this._suggestionsBoxController.widgetMounted = false;
     WidgetsBinding.instance.removeObserver(this);
+    _resizeOnScrollTimer?.cancel();
+    _effectiveFocusNode.dispose();
     super.dispose();
   }
 
@@ -686,7 +693,7 @@ class _TypeAheadFieldState<T> extends State<TypeAheadField<T>>
     }
 
     this._suggestionsBoxController =
-        _SuggestionsBoxController(context, widget.direction);
+        _SuggestionsBoxController(context, _effectiveFocusNode, widget.direction);
 
     WidgetsBinding.instance.addPostFrameCallback((duration) async {
       await this._initOverlayEntry();
@@ -704,6 +711,25 @@ class _TypeAheadFieldState<T> extends State<TypeAheadField<T>>
       // in case we already missed the focus event
       if (this._effectiveFocusNode.hasFocus) {
         this._suggestionsBoxController.open();
+      }
+
+      ScrollableState scrollableState = Scrollable.of(context);
+      if (scrollableState != null) {
+        // The TypeAheadField is inside a scrollable widget
+        scrollableState.position.isScrollingNotifier.addListener(() {
+          bool isScrolling = scrollableState.position.isScrollingNotifier.value;
+          _resizeOnScrollTimer?.cancel();
+          if (isScrolling) {
+            // Scroll started
+            _resizeOnScrollTimer =
+                Timer.periodic(_resizeOnScrollRefreshRate, (timer) {
+              _suggestionsBoxController.resize();
+            });
+          } else {
+            // Scroll finished
+            _suggestionsBoxController.resize();
+          }
+        });
       }
     });
   }
@@ -895,7 +921,7 @@ class _SuggestionsListState<T> extends State<_SuggestionsList<T>>
       this._error = null;
     });
 
-    var suggestions = [];
+    List<T> suggestions = [];
     Object error;
 
     final Object callbackIdentity = Object();
@@ -1324,6 +1350,7 @@ class _SuggestionsBoxController {
   static const int waitMetricsTimeoutMillis = 1000;
 
   final BuildContext context;
+  final FocusNode focusNode;
   final AxisDirection direction;
 
   OverlayEntry _overlayEntry;
@@ -1332,7 +1359,7 @@ class _SuggestionsBoxController {
   bool widgetMounted = true;
   double maxHeight = defaultHeight;
 
-  _SuggestionsBoxController(this.context, this.direction);
+  _SuggestionsBoxController(this.context, this.focusNode, this.direction);
 
   open() {
     if (this._isOpened) return;
@@ -1451,9 +1478,17 @@ class _SuggestionsBoxController {
     }
   }
 
+  bool _keyboardClosed() {
+    return MediaQuery.of(context).viewInsets.bottom <= 0;
+  }
+
   Future<void> onChangeMetrics() async {
     if (await _waitChangeMetrics()) {
       resize();
+    }
+    // hide the suggestions box if keyboard is hidden
+    if (widgetMounted && _keyboardClosed()) {
+      focusNode.unfocus();
     }
   }
 }